requirements.txt
numpy==1.24.3
packaging==23.1
pandas==2.0.1
pdf2image==1.16.3
pdftotext==2.2.2
Pillow==9.5.0
pytesseract==0.3.10
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
tqdm==4.65.0
tzdata==2023.3
